requirements.txt
numpy
tensorflow-cpu
scikit-learn
opencv-python-headless